scripts/requirements-shoe-edit.txt
rembg==2.0.56
pillow>=10.0.0
onnxruntime>=1.16.0
numpy>=1.24.0
